+++ package.json
@@ -9,7 +9,7 @@
     "jsdoc": "jsdoc packages/**/index.js",
     "lint": "semistandard | snazzy",
     "test:node": "node test-helper.js",
-    "test:browser": "browserify packages/**/test/index.js -t [ babelify --presets [ @babel/preset-env ] ] --debug | tape-run | faucet",
+    "test:browser": "browserify packages/**/test/*.js -t [ babelify --presets [ @babel/preset-env ] ] --debug | tape-run | faucet",
     "test": "npm run lint && npm run test:browser && npm run test:node",
     "release": "./release.sh"
   },
