+++ packages/spatial/test/mock/geojson.js
